webkish basicly complete
