+++ base/file_writer.cpp
@@ -6,19 +6,6 @@
 
 #include "file_writer.h"
 
-// modes为"ae"表示append，O_CLOEXE (close-on-exec)
-// 具体见 man 2 open
-// O_CLOEXE条目
-kish::file_writer::file_writer(const char *filename) : fptr(::fopen(filename, "ae")) {}
-
-kish::file_writer::~file_writer() {
-    ::fclose(fptr);
-}
-
-void kish::file_writer::flush() {
-    ::fflush(fptr);
-}
-
 size_t kish::file_writer::append(const char *line, size_t len) {
     size_t written = ::fwrite_unlocked(line, 1, len, fptr);
     size_t remain = len - written;
